+++ src/Player.js
@@ -14,10 +14,17 @@ const RCTAudioPlayer = NativeModules.AudioPlayer;
 
 let playerId = 0;
 
+export const PlaybackCategories = {
+  Playback: 1,
+  Ambient: 2,
+  SoloAmbient: 3,
+}
+
 const defaultPlayerOptions = {
   autoDestroy: true,
   continuesToPlayInBackground: false,
   mixWithOthers: false,
+  category: PlaybackCategories.Playback,
 };
 
 /**
@@ -40,6 +47,8 @@ class Player extends EventEmitter {
         options.continuesToPlayInBackground = defaultPlayerOptions.continuesToPlayInBackground;
       if (options.mixWithOthers == null)
         options.mixWithOthers = defaultPlayerOptions.mixWithOthers;
+      if (options.category == null)
+        options.category = defaultPlayerOptions.category;
 
       this._options = options;
     }