+++ typings/index.d.ts
@@ -14,6 +14,12 @@ declare enum MediaStates {
     PAUSED = 5
 }
 
+declare enum PlaybackCategories {
+    Playback = 1,
+    Ambient = 2,
+    SoloAmbient = 3,
+}
+
 interface BaseError<T> {
     err: "invalidpath" | "preparefail" | "startfail" | "notfound" | "stopfail" | T;
     message: string;
@@ -46,12 +52,19 @@ interface PlayerOptions {
      */
 
     continuesToPlayInBackground?: boolean;
+    
     /**
      * Boolean to determine whether other audio sources on the device will mix
      * with sounds being played back by this module.
      * (Default: false)
      */
     mixWithOthers?: boolean;
+
+    /**
+     * (iOS only) Define the audio session category
+     * (Default: Playback)
+     */
+    category?: PlaybackCategories;
 }
 
 /**